changed error checking
If there was a delay longer than 1 second in the browser, then the code would run before the page showing "no data available" would load. This could cause the code to check for the RA box before the submit request went through, so that on the next input there would not be a box available. I adjusted this so that the browser waits for 1 second before checking, and then uses a TimeoutException rather than a NoSuchElementException